e2e_tests/requirements.txt
pytest>=7.0.0
pytest-playwright>=0.4.0